# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Database API
  version: 1.0.0
  description: |
    # Knowledge Database API
    
    Comprehensive knowledge management system with AI-powered search and bilingual support.
    
    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (viewer, editor, admin)
    - Multi-language support (English and Korean)
    - AI-powered semantic search using OpenSearch
    - Real-time analytics and metrics
    - Rate limiting and request throttling
    - Comprehensive audit logging
    
    ## Authentication
    Most endpoints require authentication using JWT bearer tokens.
    Include the token in the Authorization header: `Bearer <token>`
    
    ## Rate Limiting
    - Default: 100 requests per 60 seconds per user
    - Burst: Maximum 10 requests per second
    
    ## Language Support
    Use the `language` query parameter to specify language preference:
    - `en` for English (default)
    - `ko` for Korean
  contact:
    email: api-support@knowledge.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.knowledge.example.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Development server
...
paths:
...
  /knowledge:
    get:
      summary: List knowledge items
      tags: [Knowledge Items]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/LanguageParam'
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by category ID
        - name: type
          in: query
          schema:
            type: string
            enum: [article, faq, guide, tutorial, reference]
          description: Filter by content type
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived, deleted]
          description: Filter by status (editors only)
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter by tags (must have all)
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, title, views, helpful]
            default: updated_at
          description: Sort field
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: List of knowledge items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeItemListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create knowledge item
      tags: [Knowledge Items]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeItemCreate'
      responses:
        '201':
          description: Knowledge item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeItemResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Conflict - slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /knowledge/{id}:
    get:
      summary: Get knowledge item by ID
      tags: [Knowledge Items]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/LanguageParam'
        - name: include_related
          in: query
          schema:
            type: boolean
            default: false
          description: Include related items
      responses:
        '200':
          description: Knowledge item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeItemDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update knowledge item
      tags: [Knowledge Items]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeItemUpdate'
      responses:
        '200':
          description: Knowledge item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeItemResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: Delete knowledge item
      tags: [Knowledge Items]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Knowledge item deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /knowledge/{id}/publish:
    post:
      summary: Publish knowledge item
      tags: [Knowledge Items]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Knowledge item published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeItemResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /knowledge/{id}/versions:
    get:
      summary: Get knowledge item version history
      tags: [Knowledge Items]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Version history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeVersionResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /categories:
    get:
      summary: List categories
      tags: [Categories]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LanguageParam'
        - name: parent_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by parent category
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create category
      tags: [Categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Conflict - slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /categories/{id}:
    get:
      summary: Get category details
      tags: [Categories]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/LanguageParam'
      responses:
        '200':
          description: Category details with children
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource ID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    LanguageParam:
      name: language
      in: query
      schema:
        type: string
        enum: [en, ko]
        default: en
      description: Language preference
  schemas:
...
    KnowledgeItemCreate:
      type: object
      required:
        - type
        - slug
        - title_ko
        - title_en
        - content_ko
        - content_en
      properties:
        type:
          type: string
          enum: [article, faq, guide, tutorial, reference]
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        category_id:
          type: string
          format: uuid
        title_ko:
          type: string
          maxLength: 500
        title_en:
          type: string
          maxLength: 500
        content_ko:
          type: string
        content_en:
          type: string
        summary_ko:
          type: string
        summary_en:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
        seo_title_ko:
          type: string
          maxLength: 255
        seo_title_en:
          type: string
          maxLength: 255
        seo_description_ko:
          type: string
        seo_description_en:
          type: string
        seo_keywords:
          type: array
          items:
            type: string
    KnowledgeItemUpdate:
      type: object
      properties:
        category_id:
          type: string
          format: uuid
        title_ko:
          type: string
          maxLength: 500
        title_en:
          type: string
          maxLength: 500
        content_ko:
          type: string
        content_en:
          type: string
        summary_ko:
          type: string
        summary_en:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
        seo_title_ko:
          type: string
          maxLength: 255
        seo_title_en:
          type: string
          maxLength: 255
        seo_description_ko:
          type: string
        seo_description_en:
          type: string
        seo_keywords:
          type: array
          items:
            type: string
    KnowledgeItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        slug:
          type: string
        category:
          $ref: '#/components/schemas/CategoryInfo'
        title_ko:
          type: string
        title_en:
          type: string
        summary_ko:
          type: string
        summary_en:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published, archived, deleted]
        version:
          type: integer
        view_count:
          type: integer
        helpful_count:
          type: integer
        not_helpful_count:
          type: integer
        published_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    KnowledgeItemDetailResponse:
      allOf:
        - $ref: '#/components/schemas/KnowledgeItemResponse'
        - type: object
          properties:
            content_ko:
              type: string
            content_en:
              type: string
            metadata:
              type: object
            seo_title_ko:
              type: string
            seo_title_en:
              type: string
            seo_description_ko:
              type: string
            seo_description_en:
              type: string
            seo_keywords:
              type: array
              items:
                type: string
            related_items:
              type: array
              items:
                $ref: '#/components/schemas/KnowledgeItemResponse'
    KnowledgeItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeItemResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean
    KnowledgeVersionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version_number:
          type: integer
        title_ko:
          type: string
        title_en:
          type: string
        change_summary:
          type: string
        created_at:
          type: string
          format: date-time
    CategoryInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name_ko:
          type: string
        name_en:
          type: string
        slug:
          type: string
    CategoryCreate:
      type: object
      required:
        - name_ko
        - name_en
        - slug
      properties:
        parent_id:
          type: string
          format: uuid
        name_ko:
          type: string
        name_en:
          type: string
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        description_ko:
          type: string
        description_en:
          type: string
        icon:
          type: string
        display_order:
          type: integer
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        display_order:
          type: integer
        item_count:
          type: integer
    CategoryDetailResponse:
      allOf:
        - $ref: '#/components/schemas/CategoryResponse'
        - type: object
          properties:
            children:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  slug:
                    type: string
                  item_count:
                    type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: unauthorized
            message: Invalid or expired token
    ForbiddenError:
      description: User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: forbidden
            message: You don't have permission to access this resource
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: not_found
            message: Resource not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string
          example:
            error: validation_error
            message: Validation failed
            details:
              - loc: ["body", "email"]
                msg: invalid email address
                type: value_error.email
...